+++ packages/spacecat-shared-utils/src/index.js
@@ -61,3 +61,4 @@ export { s3Wrapper } from './s3.js';
 
 export { fetch } from './adobe-fetch.js';
 export { tracingFetch } from './tracing-fetch.js';
+export { getHighFormViewsLowConversionMetrics } from './formcalc.js';